Group Activity task scoring is influenced by instigator
Changes:
- GroupActivityMeta impleemnts specific logic for Person suitability
	- If person is the instigator; then high score
	- If person is has a relationship opinion on the inmstigator; use it to create a modifier
- Relation defines the max opinion and uses MathUtils

Part of #971

+++ mars-sim-core/src/test/java/com/mars_sim/core/AbstractMarsSimUnitTest.java
@@ -229,7 +229,7 @@ public abstract class AbstractMarsSimUnitTest extends TestCase {
 	/**
 	 * Better Assert method 
 	 */
-	public static void assertGreaterThan(String message, int minValue, int actual) {
+	public static void assertGreaterThan(String message, double minValue, double actual) {
 		if (actual < minValue) {
 			fail(message + " ==> " +
 					"Expected: a value greater than <" + minValue + ">\n" +
